image_alignment/GD/requirements.txt
matplotlib==3.9.2
numpy==2.1.1
Pillow==10.4.0
skimage==0.0
torch==2.4.0
torchvision==0.19.0
